minor corrections

+++ server/controllers/accomodation.server.controller.js
@@ -105,15 +105,18 @@ exports.delete = function (req, res, next) {
         if (err) {
             return next(err);
         }
-        const updir = `${dir}/${accomodation._id}`;
-        accomodation.photos.forEach(photo => {
-            fs.unlinkSync(`${updir}/${photo}`);
-        });
-        fs.rmdir(updir, (err) => {
-            if (err) {
-                return next(err);
-            }
-        });
+        if (accomodation.photos.length) {
+            const updir = `${dir}/${accomodation._id}`;
+            accomodation.photos.forEach(photo => {
+                fs.unlinkSync(`${updir}/${photo}`);
+            });
+            fs.rmdir(updir, (err) => {
+                if (err) {
+                    return next(err);
+                }
+            });
+        }
+
         res.status(200).json(accomodation);
     })
 };
